add -Og
The build failure is due to memory consumption. This seems to be the
only change between `feature/library-ga` and earlier succeeding builds.
If this is the issue, I'm going to try to compensate this. For the test
builds, `-Og` isn't an issue, and should be set using `CFLAGS` to
speedup builds, but it might be nice to enable some optimisations for
production builds.